[TASK] Move StandardVariableProviderModelFixture to unit namespace
Move this fixture file to a more appropriate place.
